Finished all other endpoints and TODOs. Corrected the guess and answer comparison on the frontend.

+++ projects/02_trivia_api/starter/backend/flaskr/__init__.py
@@ -9,7 +9,9 @@ from models import setup_db, Question, Category
 
 QUESTIONS_PER_PAGE = 10
 
-def get_paginated_results(request, data):
+def paginate(request, data):
+  if not data:
+    return data
   page = request.args.get('page', 1, type=int)
   start = (page - 1) * QUESTIONS_PER_PAGE
   end = start + QUESTIONS_PER_PAGE
@@ -44,12 +46,12 @@ def create_app(test_config=None):
     try:
       categories = Category.query.all()
       formatted_categories = {category.id: category.type for category in categories}
+      return jsonify({
+        'success': True,
+        'categories': formatted_categories
+      })
     except:
       abort(404)
-    return jsonify({
-      'success': True,
-      'categories': formatted_categories
-    })
 
   '''
   @TODO: 
@@ -67,19 +69,19 @@ def create_app(test_config=None):
   def get_questions():
     try:
       questions = Question.query.all()
-      paginated_qs = get_paginated_results(request, questions)
+      paginated_qs = paginate(request, questions)
 
       categories = Category.query.all()
       formatted_categories = {category.id: category.type for category in categories}
+      return jsonify({
+        'success': True,
+        'questions': [q.format() for q in paginated_qs],
+        'total_questions': len(questions),
+        'categories': formatted_categories,
+        'current_category': paginated_qs[0].category
+      })
     except:
       abort(404)
-    return jsonify({
-      'success': True,
-      'questions': [q.format() for q in paginated_qs],
-      'total_questions': len(questions),
-      'categories': formatted_categories,
-      'current_category': paginated_qs[0].category
-    })
 
   '''
   @TODO: 
@@ -92,12 +94,15 @@ def create_app(test_config=None):
   def delete_question(q_id):
     try:
       question = Question.query.get(q_id)
+      if not question:
+        abort(404)
+
       question.delete()
+      return jsonify({
+        'success': True
+      })
     except:
-      abort(404)
-    return jsonify({
-      'success': True
-    })
+      abort(422)
 
   '''
   @TODO: 
@@ -109,17 +114,6 @@ def create_app(test_config=None):
   the form will clear and the question will appear at the end of the last page
   of the questions list in the "List" tab.  
   '''
-  @app.route('/questions', methods=['POST'])
-  def create_question():
-    body = json.loads(request.data)
-    try:
-      question = Question(question=body['question'], answer=body['answer'], category=body['category'], difficulty=body['difficulty'])
-      question.insert()
-    except:
-      abort(400)
-    return jsonify({
-      'success': True
-    })
 
   '''
   @TODO: 
@@ -131,6 +125,36 @@ def create_app(test_config=None):
   only question that include that string within their question. 
   Try using the word "title" to start. 
   '''
+  @app.route('/questions', methods=['POST'])
+  def create_or_search_question():
+    body = json.loads(request.data)
+    new_question = body.get('question', None)
+    new_answer = body.get('answer', None)
+    new_category = body.get('category', None)
+    new_difficulty = body.get('difficulty', None)
+    search = body.get('searchTerm', None)
+
+    try:
+      if search:
+        results = Question.query.filter(Question.question.ilike('%{}%'.format(search))).all()
+        questions = paginate(request, results)
+        formatted_qs = [q.format() for q in questions]
+        
+        return jsonify({
+          'success': True,
+          'questions': formatted_qs,
+          'total_questions': len(results),
+          'current_category': questions[0].category
+        })
+      else:
+        question = Question(question=new_question, answer=new_answer, category=new_category, difficulty=new_difficulty)
+        question.insert()
+        return jsonify({
+          'success': True
+        })
+    except Exception as e:
+      print('Error: ', e)
+      abort(400)
 
   '''
   @TODO: 
@@ -140,26 +164,83 @@ def create_app(test_config=None):
   categories in the left column will cause only questions of that 
   category to be shown. 
   '''
-
+  @app.route('/categories/<int:category_id>/questions')
+  def get_questions_for_category(category_id):
+    try:
+      results = Question.query.filter(Question.category == category_id).all()
+      questions = paginate(request, results)
+      formatted_qs = [q.format() for q in questions]
+      return jsonify({
+        'success': True,
+        'questions': formatted_qs,
+        'total_questions': len(results),
+        'current_category': questions[0].category
+      })
+    except:
+      abort(404)
 
   '''
   @TODO: 
   Create a POST endpoint to get questions to play the quiz. 
   This endpoint should take category and previous question parameters 
-  and return a random questions within the given category, 
+  and return a random question within the given category, 
   if provided, and that is not one of the previous questions. 
 
   TEST: In the "Play" tab, after a user selects "All" or a category,
   one question at a time is displayed, the user is allowed to answer
   and shown whether they were correct or not. 
   '''
+  @app.route('/quizzes', methods=['POST'])
+  def get_quiz_question():
+    body = json.loads(request.data)
+    previous = body.get('previous_questions', [])
+    category = body.get('quiz_category', None)
+
+    try:
+      if category:
+        category_id = category.get('id', 0)
+        questions = Question.query.filter(Question.id.notin_(previous), Question.category == category_id).all()
+      else:
+        questions = Question.query.filter(Question.id.notin_(previous)).all()
+      
+      question = random.choice(questions).format() if questions else None
+      return jsonify({
+          'success': True,
+          'question': question,
+      })
+    except Exception as e:
+      print('Error:', e)
+      abort(404)
 
   '''
   @TODO: 
   Create error handlers for all expected errors 
   including 404 and 422. 
   '''
+  @app.errorhandler(400)
+  def bad_request():
+    return jsonify({
+      'succes': False,
+      'error': 400,
+      'message': 'Request was not formatted correctly.'
+    }), 400
+
+  @app.errorhandler(404)
+  def not_found():
+    return jsonify({
+      'success': False,
+      'error': 404
+      'message': 'Question not found.'
+    }), 404
   
+  @app.errorhandler(422)
+  def unprocessable():
+    return jsonify({
+      'success': False,
+      'error': 422,
+      'message': 'Data within the request body was unable to be processed.'
+    }), 422
+
   return app
